Created a LICENSE file containing the MIT License text.
Added the MIT License to the project to clarify the terms under which the code can be used, modified, and distributed. The MIT License is a permissive open-source license that allows users to freely use, modify, and distribute the software for any purpose, provided that the copyright notice and permission notice are included. This license was chosen to encourage collaboration and adoption of the project by other developers and organizations.